requirements.txt
flask
flask-cors
rembg
pillow
waitress
numpy
onnxruntime
onnxruntime-tools
gunicorn

# Flask==3.0.3
# flask-cors==4.0.1
# gunicorn==21.2.0
# Pillow==10.4.0
# rembg==2.0.57
# torch==2.1.0
# torchvision==0.16.0